# legacy/v1/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Secure MCP Server API
  description: Enterprise-grade Model Context Protocol server with advanced security and scalability
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@enterprise.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.secure-mcp.enterprise.com/api/v1
    description: Production server
  - url: https://staging.secure-mcp.enterprise.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required:
        - error
        - message
        - timestamp
        - requestId
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid request parameters
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          example: req_abc123
        traceId:
          type: string
          example: trace_xyz789
    User:
      type: object
      properties:
        id:
          type: string
          example: user_123
        email:
          type: string
          format: email
        name:
          type: string
        roles:
          type: array
          items:
            type: string
            enum: [admin, developer, user]
        permissions:
          type: array
          items:
            type: string
        mfaEnabled:
          type: boolean
        status:
          type: string
          enum: [active, suspended, deleted]
        lastLogin:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Session:
      type: object
      properties:
        id:
          type: string
          example: session_123
        userId:
          type: string
        contextId:
          type: string
        status:
          type: string
          enum: [active, terminated]
        connectionType:
          type: string
          enum: [websocket, http]
        ipAddress:
          type: string
        userAgent:
          type: string
        startedAt:
          type: string
          format: date-time
        lastActivity:
          type: string
          format: date-time
        metrics:
          type: object
          properties:
            messagesReceived:
              type: integer
            messagesSent:
              type: integer
            toolExecutions:
              type: integer
            errors:
              type: integer
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimitError:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                mfaCode:
                  type: string
                  pattern: '^[0-9]{6}$'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  expiresIn:
                    type: integer
                  tokenType:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
                  tokenType:
                    type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /auth/mfa/verify:
    post:
      tags:
        - Authentication
      summary: Verify MFA code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - code
                - type
              properties:
                code:
                  type: string
                type:
                  type: string
                  enum: [totp, sms]
      responses:
        '200':
          description: MFA verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                  timestamp:
                    type: string
                    format: date-time
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /sessions:
    get:
      tags:
        - Sessions
      summary: List sessions
      responses:
        '200':
          description: Sessions list
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...